不要な型変換を行わない仕様に変更。

+++ Implem.Pleasanter/Libraries/DataTypes/ProgressRate.cs
@@ -121,10 +121,10 @@ namespace Implem.Pleasanter.Libraries.DataTypes
                 : CreatedTime;
         }
 
-        private float PlannedValue(DateTime now, DateTime start, double range)
+        private double PlannedValue(DateTime now, DateTime start, double range)
         {
             return start < now && range > 0
-                ? (float)(Elapsed(now, start)) / (float)range
+                ? (Elapsed(now, start)) / range
                 : 0;
         }
